static LOD force hook cleanup, added a new better config getter function
didn't replace all of them, probably won't do it myself

+++ xlive/Blam/Engine/render/render_lod_new.h
@@ -2,7 +2,7 @@
 
 #include "memory/static_arrays.h"
 
-enum e_render_lod : uint32
+enum e_render_lod : int8
 {
 	_render_lod_disabled = 0,
 	_render_lod_super_low = 1,
