+++ xlive/H2MOD/Modules/Shell/Config.h
@@ -60,7 +60,7 @@ extern bool H2Config_raw_input;
 extern bool H2Config_discord_enable;
 extern bool H2Config_controller_aim_assist;
 extern int H2Config_fps_limit;
-extern int H2Config_static_lod_state;
+extern int8 H2Config_static_lod_state;
 extern int H2Config_field_of_view;
 extern int H2Config_vehicle_field_of_view;
 extern bool H2Config_static_first_person;
